Added Files, Handlebars, Styles, Fixed Routing
Added a bunch of new files, fixed routing code, changed styling, added more handlebars, migrated the old code, added client-side code, working on making the dataTable dynamic with our data, working on new styling for a beautiful table.

Fixed forms to send data correctly, got the data fetching API to work.

Made a helper function to create compact objects for dynamic form rendering. Done as a proof of concept for dynamic table creation using complex objects. TODO: Make object creation functions for each of the tables for easy data display.

Still need to set up the server database on the OSU servers, and I gotta connect the site to that too.

Also, still need to make the deletion/update events as well. As soon as the table is developed I can spend a lot of time on that and get that done.